refactor: reorganize codebase structure and extract large components
- Extract BookList (388 lines) into smaller components:
  - BookCard: individual book display
  - BookProgress: progress and rating functionality
  - BookActions: action buttons and context menu
  - BookList: container and state management

- Reorganize components into logical folders:
  - components/book/: Book-related components
  - components/forms/: Form components
  - components/navigation/: Navigation components
  - components/statistics/: Statistics components
  - components/filters/: Filter components
  - components/ui/: Shared UI components

- Move theme configuration to dedicated folder:
  - src/theme/: Theme config and provider

- Update all import paths and fix TypeScript errors
- Add barrel exports for clean imports
- Improve maintainability and code organization

Benefits:
- Better maintainability with smaller, focused components
- Improved readability and separation of concerns
- Easier testing and reusability
- Cleaner import statements
- Better scalability for future development
