+++ src/App.tsx
@@ -1,8 +1,6 @@
 import { Outlet } from "react-router-dom";
 
-import Navbar from "./components/Navbar";
-import MobileNavbar from "./components/MobileNavbar";
-import ErrorBoundary from "./components/ErrorBoundary";
+import { Navbar, MobileNavbar, ErrorBoundary } from "./components";
 import "./App.css";
 
 function App() {
