+++ src/components/book/index.ts
@@ -1,0 +1,4 @@
+export { default as BookCard } from './BookCard';
+export { default as BookList } from './BookList';
+export { default as BookProgress } from './BookProgress';
+export { default as BookActions } from './BookActions';
