+++ src/components/forms/index.ts
@@ -1,0 +1,2 @@
+export { default as BookForm } from './AddBookForm/BookForm';
+export { default as BookForms } from './BookForms';
