+++ src/components/index.ts
@@ -1,20 +1,18 @@
 // Barrel exports for components
-export { default as BookForm } from "./AddBookForm/BookForm";
-export { default as BookForms } from "./BookForms";
-export { default as BookList } from "./BookList";
-export { default as BottomNav } from "./BottomNav";
-export { default as ErrorBoundary } from "./ErrorBoundary";
-export { default as ErrorDisplay } from "./ErrorDisplay";
-export { default as FilterSortPanel } from "./FilterSortPanel";
-export { default as Input } from "./Input";
-export { default as LoadingSpinner } from "./LoadingSpinner";
-export { default as MetricCard } from "./MetricCard";
-export { default as MetricsGrid } from "./MetricsGrid";
-export { default as MobileNavbar } from "./MobileNavbar";
-export { default as Navbar } from "./Navbar";
-export { default as PageHeader } from "./PageHeader";
-export { default as ProtectedRoute } from "./ProtectedRoute";
-export { default as StatCard } from "./StatCard";
-export { default as StatisticsDashboard } from "./StatisticsDashboard";
-export { default as StatisticsGrid } from "./StatisticsGrid";
-export { default as StatisticsHeader } from "./StatisticsHeader";
+// Book components
+export * from './book';
+
+// Form components
+export * from './forms';
+
+// Navigation components
+export * from './navigation';
+
+// Statistics components
+export * from './statistics';
+
+// Filter components
+export * from './filters';
+
+// UI components
+export * from './ui';
