+++ src/components/ui/index.ts
@@ -1,0 +1,6 @@
+export { default as ErrorBoundary } from './ErrorBoundary';
+export { default as ErrorDisplay } from './ErrorDisplay';
+export { default as LoadingSpinner } from './LoadingSpinner';
+export { default as Input } from './Input';
+export { default as PageHeader } from './PageHeader';
+export { default as ProtectedRoute } from './ProtectedRoute';
