+++ src/theme/index.ts
@@ -1,0 +1,2 @@
+export { default as ThemeProvider } from './ThemeProvider';
+export { default as theme } from './theme';
